refactor(explore): isolate stage document delivery
